+++ src/ScottPlot5/ScottPlot5/Plottables/Crosshair.cs
@@ -1,6 +1,6 @@
 ﻿namespace ScottPlot.Plottables;
 
-public class Crosshair : IPlottable, IRenderLast
+public class Crosshair : IPlottable, IRenderLast, IHasMarker
 {
     public HorizontalLine HorizontalLine { get; } = new();
     public VerticalLine VerticalLine { get; } = new();
@@ -18,6 +18,15 @@ public class Crosshair : IPlottable, IRenderLast
     public Color LineColor { set { HorizontalLine.LineColor = value; VerticalLine.LineColor = value; } }
     public float LineWidth { set { HorizontalLine.LineWidth = value; VerticalLine.LineWidth = value; } }
     public LinePattern LinePattern { set { HorizontalLine.LinePattern = value; VerticalLine.LinePattern = value; } }
+
+    public MarkerStyle MarkerStyle { get; set; } = new() { LineWidth = 1 };
+    public MarkerShape MarkerShape { get => MarkerStyle.Shape; set => MarkerStyle.Shape = value; }
+    public float MarkerSize { get => MarkerStyle.Size; set => MarkerStyle.Size = value; }
+    public Color MarkerFillColor { get => MarkerStyle.FillColor; set => MarkerStyle.FillColor = value; }
+    public Color MarkerLineColor { get => MarkerStyle.LineColor; set => MarkerStyle.LineColor = value; }
+    public Color MarkerColor { get => MarkerStyle.MarkerColor; set => MarkerStyle.MarkerColor = value; }
+    public float MarkerLineWidth { get => MarkerStyle.LineWidth; set => MarkerStyle.LineWidth = value; }
+
 
     [Obsolete("Assign values to LineColor, LineWidth, or LinePattern properties to set style for both lines. " +
         "HorizontalLine and VerticalLine have properties that can be set individually as well.", true)]
@@ -53,5 +62,9 @@ public class Crosshair : IPlottable, IRenderLast
 
         VerticalLine.Axes = Axes;
         VerticalLine.RenderLast(rp);
+
+        using SKPaint paint = new();
+        Pixel px = Axes.GetPixel(new Coordinates(X, Y));
+        MarkerStyle.Render(rp.Canvas, px, paint);
     }
 }